日志配置收口&文档update
